requirements.txt
aioredis==1.3.1
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.3
django-celery-results==2.3.0
django-cors-headers==3.10.0
django-extensions==3.1.3
django-picklefield==3.0.1
djangorestframework==3.12.4
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
joblib==1.1.0
kombu==5.2.4
msgpack==1.0.3
msgpack-python==0.5.6
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==8.3.2
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.2.0
pyngrok==5.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==5.4.1
redis==4.1.4
requests==2.26.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
twilio==7.1.0
Twisted==22.2.0
txaio==22.2.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
zope.interface==5.4.0
